services/chat/tests/requirements-test.txt
pytest
requests
python-dotenv
pytest-cov
